feat: option type
