+++ src/lib/component/Triangle.tsx
@@ -1,23 +1,4 @@
-// import "../../example/test.css";
-
-import { ReactNode } from "react";
-
-type StringObject = {
-  [key: string]: string;
-};
-
-interface PropsType {
-  option: {
-    width: string;
-    height: string;
-    borderRadius?: string;
-    color?: string;
-  };
-  className?: string;
-  style?: StringObject;
-  children?: ReactNode;
-  id?: string;
-}
+import { PropsType } from "../util/type";
 
 export default function Triangle({
   id,
